+++ cdp-agentkit-core/python/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Added `get_balance_nft` action.
+- Added `transfer_nft` action.
+
 ## [0.0.8] - 2025-01-13
 
 ### Added
@@ -13,6 +18,11 @@
 ### Added
 
 - Bump `cdp-sdk` dependency to `0.13.0`.
+
+### Added
+
+- Added `get_balance_nft` action.
+- Added `transfer_nft` action.
 
 ## [0.0.6] - 2024-12-06